# add uid cache for UID of a mailbox (folder) to improve performance